linkedin_server/requirements.txt
# MCP SDK
mcp>=0.1.0

# HTTP requests
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0